+++ ChangeLog.md
@@ -4,3 +4,4 @@
 * Add: NEIU Initial Compatibility (NYI)
 * Add: Waila Compatibility (NYI)
 * Add: /emi command
+* Change: Optimized recipe IDs